Renamed link-if to next-if
