+++ src/index.ts
@@ -1,7 +1,7 @@
 export { default as cloneState } from './clone-state';
 export { default as compose } from './compose';
-export { default as linkIf, default as filter } from './link-if';
-export { default as linkIfType, default as filterIfType } from './link-if-type';
+export { default as nextIf } from './next-if';
+export { default as nextIfType } from './next-if-type';
 export { default as mergeStates } from './merge-states';
 export { default as mapByType } from './map-by-type';
 export { default as withDefault } from './with-default';
