+++ src/next-if.ts
@@ -1,10 +1,10 @@
 import { ComposableReducer } from "./types";
 
-const linkIf = <TState, TAction>(
+const nextIf = <TState, TAction>(
   predicate: (state: TState, action: TAction) => boolean
 ): ComposableReducer<TState, TAction> => {
   return (state: TState, action: TAction, next?: (state: TState) => TState) =>
     !!next && predicate(state, action) ? next(state) : state;
 };
 
-export default linkIf;
+export default nextIf;
