+++ test/next-if.spec.ts
@@ -1,22 +1,19 @@
-import { filter, linkIf } from "../src";
+import { nextIf } from "../src";
 
 describe("filter", () => {
   it("is exported properly", () => {
-    expect(typeof filter).toBe("function");
+    expect(typeof nextIf).toBe("function");
   });
 
-  it("has the linkIf alias", () => {
-    expect(filter).toEqual(linkIf);
-  });
   it('calls "next" if the predicate returns truthy', () => {
-    const reducer = filter(() => true);
+    const reducer = nextIf(() => true);
     const next = jest.fn();
     reducer({}, {}, next);
     expect(next).toHaveBeenCalledTimes(1);
   });
 
   it('doesn\'t call "next" if the predicate returns falsy', () => {
-    const reducer = filter(() => false);
+    const reducer = nextIf(() => false);
     const next = jest.fn();
     reducer({}, {}, next);
     expect(next).not.toHaveBeenCalled();
